Added CubemapProbe::setPriority When using per pixel reflections, cubemap areas were relaxed to only intersect with the probe shape, rather than being contained inside it. This gives slightly more flexibility when ensuring blending transition between works smoothly

+++ Components/Hlms/Pbs/include/Cubemaps/OgreParallaxCorrectedCubemapAuto.h
@@ -93,7 +93,7 @@ namespace Ogre
                                       const CompositorWorkspaceDef *probeWorkspaceDef );
         ~ParallaxCorrectedCubemapAuto();
 
-        virtual TextureGpu* _acquireTextureSlot( uint32 &outTexSlot );
+        virtual TextureGpu* _acquireTextureSlot( uint16 &outTexSlot );
         virtual void _releaseTextureSlot( TextureGpu *texture, uint32 texSlot );
 
         virtual TextureGpu* findTmpRtt( const TextureGpu *baseParams );
